- Move EndShipper to GA - EndShipper all returns collection rather than list
