+++ EasyPost.Tests/cassettes/netstandard/end_shipper/all.json
@@ -1,7 +1,7 @@
 [
     {
-        "Duration": 241,
-        "RecordedAt": "2022-09-13T13:55:18.006646-06:00",
+        "Duration": 236,
+        "RecordedAt": "2022-09-16T12:16:52.782113-06:00",
         "Request": {
             "Body": "",
             "BodyContentType": "Text",
@@ -13,15 +13,15 @@
                 "Accept": "application/json,text/json,text/x-json,text/javascript,application/xml,text/xml",
                 "User-Agent": "<REDACTED>"
             },
-            "Uri": "https://api.easypost.com/beta/end_shippers?page_size=5"
+            "Uri": "https://api.easypost.com/v2/end_shippers?page_size=5"
         },
         "Response": {
-            "Body": "[{\"id\":\"es_1c5d0739799f408e9943383c639b30f2\",\"object\":\"EndShipper\",\"created_at\":\"2022-05-17T17:43:19-06:00\",\"updated_at\":\"2022-05-17T17:43:19-06:00\",\"name\":\"JACK SPARROW\",\"company\":\"EASYPOST\",\"street1\":\"388 TOWNSEND ST APT 20\",\"street2\":null,\"city\":\"SAN FRANCISCO\",\"state\":\"CA\",\"zip\":\"94107-1670\",\"country\":\"US\",\"phone\":\"<REDACTED>\",\"email\":\"<REDACTED>\"},{\"id\":\"es_e591f790ddeb497a9469eb9104d9a90b\",\"object\":\"EndShipper\",\"created_at\":\"2022-05-17T17:43:19-06:00\",\"updated_at\":\"2022-05-17T17:43:19-06:00\",\"name\":\"JACK SPARROW\",\"company\":\"EASYPOST\",\"street1\":\"388 TOWNSEND ST APT 20\",\"street2\":null,\"city\":\"SAN FRANCISCO\",\"state\":\"CA\",\"zip\":\"94107-1670\",\"country\":\"US\",\"phone\":\"<REDACTED>\",\"email\":\"<REDACTED>\"},{\"id\":\"es_3e5d2c14d0c24434b2809058fe9bf764\",\"object\":\"EndShipper\",\"created_at\":\"2022-05-17T17:43:20-06:00\",\"updated_at\":\"2022-05-17T17:43:20-06:00\",\"name\":\"JACK SPARROW\",\"company\":\"EASYPOST\",\"street1\":\"388 TOWNSEND ST APT 20\",\"street2\":null,\"city\":\"SAN FRANCISCO\",\"state\":\"CA\",\"zip\":\"94107-1670\",\"country\":\"US\",\"phone\":\"<REDACTED>\",\"email\":\"<REDACTED>\"},{\"id\":\"es_042fc0f4620d472c888fcac88065ef2d\",\"object\":\"EndShipper\",\"created_at\":\"2022-05-18T15:55:08-06:00\",\"updated_at\":\"2022-05-18T15:55:08-06:00\",\"name\":\"JACK SPARROW\",\"company\":\"EASYPOST\",\"street1\":\"388 TOWNSEND ST APT 20\",\"street2\":null,\"city\":\"SAN FRANCISCO\",\"state\":\"CA\",\"zip\":\"94107-1670\",\"country\":\"US\",\"phone\":\"<REDACTED>\",\"email\":\"<REDACTED>\"},{\"id\":\"es_aa2adbe577ff40bb9e64c082062c0f3c\",\"object\":\"EndShipper\",\"created_at\":\"2022-05-18T15:55:09-06:00\",\"updated_at\":\"2022-05-18T15:55:09-06:00\",\"name\":\"JACK SPARROW\",\"company\":\"EASYPOST\",\"street1\":\"388 TOWNSEND ST APT 20\",\"street2\":null,\"city\":\"SAN FRANCISCO\",\"state\":\"CA\",\"zip\":\"94107-1670\",\"country\":\"US\",\"phone\":\"<REDACTED>\",\"email\":\"<REDACTED>\"}]",
+            "Body": "{\"end_shippers\":[{\"id\":\"es_1c5d0739799f408e9943383c639b30f2\",\"object\":\"EndShipper\",\"created_at\":\"2022-05-17T17:43:19-06:00\",\"updated_at\":\"2022-05-17T17:43:19-06:00\",\"name\":\"JACK SPARROW\",\"company\":\"EASYPOST\",\"street1\":\"388 TOWNSEND ST APT 20\",\"street2\":null,\"city\":\"SAN FRANCISCO\",\"state\":\"CA\",\"zip\":\"94107-1670\",\"country\":\"US\",\"phone\":\"<REDACTED>\",\"email\":\"<REDACTED>\"},{\"id\":\"es_e591f790ddeb497a9469eb9104d9a90b\",\"object\":\"EndShipper\",\"created_at\":\"2022-05-17T17:43:19-06:00\",\"updated_at\":\"2022-05-17T17:43:19-06:00\",\"name\":\"JACK SPARROW\",\"company\":\"EASYPOST\",\"street1\":\"388 TOWNSEND ST APT 20\",\"street2\":null,\"city\":\"SAN FRANCISCO\",\"state\":\"CA\",\"zip\":\"94107-1670\",\"country\":\"US\",\"phone\":\"<REDACTED>\",\"email\":\"<REDACTED>\"},{\"id\":\"es_3e5d2c14d0c24434b2809058fe9bf764\",\"object\":\"EndShipper\",\"created_at\":\"2022-05-17T17:43:20-06:00\",\"updated_at\":\"2022-05-17T17:43:20-06:00\",\"name\":\"JACK SPARROW\",\"company\":\"EASYPOST\",\"street1\":\"388 TOWNSEND ST APT 20\",\"street2\":null,\"city\":\"SAN FRANCISCO\",\"state\":\"CA\",\"zip\":\"94107-1670\",\"country\":\"US\",\"phone\":\"<REDACTED>\",\"email\":\"<REDACTED>\"},{\"id\":\"es_042fc0f4620d472c888fcac88065ef2d\",\"object\":\"EndShipper\",\"created_at\":\"2022-05-18T15:55:08-06:00\",\"updated_at\":\"2022-05-18T15:55:08-06:00\",\"name\":\"JACK SPARROW\",\"company\":\"EASYPOST\",\"street1\":\"388 TOWNSEND ST APT 20\",\"street2\":null,\"city\":\"SAN FRANCISCO\",\"state\":\"CA\",\"zip\":\"94107-1670\",\"country\":\"US\",\"phone\":\"<REDACTED>\",\"email\":\"<REDACTED>\"},{\"id\":\"es_aa2adbe577ff40bb9e64c082062c0f3c\",\"object\":\"EndShipper\",\"created_at\":\"2022-05-18T15:55:09-06:00\",\"updated_at\":\"2022-05-18T15:55:09-06:00\",\"name\":\"JACK SPARROW\",\"company\":\"EASYPOST\",\"street1\":\"388 TOWNSEND ST APT 20\",\"street2\":null,\"city\":\"SAN FRANCISCO\",\"state\":\"CA\",\"zip\":\"94107-1670\",\"country\":\"US\",\"phone\":\"<REDACTED>\",\"email\":\"<REDACTED>\"}],\"has_more\":true}",
             "BodyContentType": "Json",
             "ContentHeaders": {
                 "Expires": "0",
                 "Content-Type": "application/json; charset=utf-8",
-                "Content-Length": "1801"
+                "Content-Length": "1834"
             },
             "HttpVersion": "1.1",
             "ResponseHeaders": {
@@ -31,15 +31,15 @@
                 "x-download-options": "noopen",
                 "x-permitted-cross-domain-policies": "none",
                 "referrer-policy": "strict-origin-when-cross-origin",
-                "x-ep-request-uuid": "f5e0a5a46320e026ecb357100078af6b",
+                "x-ep-request-uuid": "c1f44a406324bd94ecd6ee4300002a2c",
                 "Cache-Control": "no-store, no-cache, private",
                 "Pragma": "no-cache",
-                "ETag": "W/\"00c862705e26113b8ffeb007f410f95f\"",
-                "x-runtime": "0.065212",
-                "x-node": "bigweb3nuq",
-                "x-version-label": "easypost-202209122308-273e11ecbb-master",
+                "ETag": "W/\"76b46380ab8526bb50fdc6cecce53fbb\"",
+                "x-runtime": "0.061214",
+                "x-node": "bigweb2nuq",
+                "x-version-label": "easypost-202209152244-6075a793a4-master",
                 "x-backend": "easypost",
-                "x-proxied": "intlb1nuq 6aa9972c1e,extlb1nuq 9ce216bfac",
+                "x-proxied": "intlb1nuq 6aa9972c1e,extlb2nuq 9ce216bfac",
                 "Strict-Transport-Security": "max-age=31536000; includeSubDomains; preload"
             },
             "Status": {
